Restructure peristence layer
