+++ src/new_model/runtime/persistence/flow_node_instance/index.ts
@@ -1,0 +1,2 @@
+export * from './repository';
+export * from './service';
