+++ src/new_model/runtime/persistence/process_model/index.ts
@@ -1,0 +1,2 @@
+export * from './repository';
+export * from './service';
