--- src/new_model/runtime/storage/index.ts
@@ -1,3 +1,0 @@
-export * from './repository/index';
-export * from './security/index';
-export * from './service/index';
