--- src/new_model/runtime/storage/repository/index.ts
@@ -1,2 +1,0 @@
-export * from './process_model_persistence_repository';
-export * from './flow_node_instance_persistence_repository';
