--- src/new_model/runtime/storage/service/index.ts
@@ -1,2 +1,0 @@
-export * from './process_model_persistence_service';
-export * from './flow_node_instance_persistence_service';
